requirements.txt
argparse
pandas
biopython
atom3d
matplotlib
seaborn
jupyterlab
scipy
sklearn
wandb